Move `ButtonsUI` implementation into `buttons_ui` module
